requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.62.1
greenlet==3.4.0
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
openpyxl==3.1.5
packaging==26.1
pandas==3.0.2
pillow==12.2.0
psycopg2-binary==2.9.11
pyparsing==3.3.2
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
shiboken6==6.7.2
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
xlsxwriter==3.2.9
